ivan: Remove oppo logical partitions when necessary
Usually when building full OTAs for dynamic partition devices the whole partition table is removed and recreated
but in vendorless cases lineageos (and basically any other custom rom) has a patch to disable that functionality and
instead just resize the partitions. This, however, breaks the flashing if a new partition is added, because that
partition is tried to be resized and, since it doesn't exist, the assert fails. This patch adds another dynamic patch
list to remove Oppo partitions or gracefully fail if it already removed (no assert).

Signed-off-by: Kuba Wojciechowski <nullbytepl@gmail.com>
Change-Id: I1f316c9f16b20324674b55b29f69dfc8c0d642a5